url: Don't pass UTF-8 hostname to libidn unless it's valid UTF-8. draft5
- Add a unit test.

There is some question about whether we should be returning a better
error (currently, the validation error, IDNA_STRINGPREP_ERROR, gets lost
because the only action we take on invalid input is to not convert
non-ascii hostnames, which eventually causes a dns failure; that's what
gets reported). The unit test for the fix thus tests dns failure instead
of looking for a specific error from utf8 validation.